wsgi.py
# -*- coding: utf-8 -*-
__author__ = '__imzqqq'

import os
from dotenv import load_dotenv
from app import create_app


dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
if os.path.exists(dotenv_path):
    load_dotenv(dotenv_path, override=True)

app = create_app()
