requirements.txt
Flask
PyMySQL
requests
flask-login
pillow
pandas
mysql.connector
flask_bootstrap
flask_login
Flask_WTF
Click
Flask_Sitemap
Werkzeug
WTForms
python-dotenv
Flask-Migrate
Flask-SQLAlchemy
xml2dict
re
datetime